Now using npm package to read csv instead of custom funciton. Also moved funciton from Generic to SlurmClusterInterface
