+++ HCCGo/app/js/GenericClusterInterface.js
@@ -17,84 +17,6 @@ GenericClusterInterface.prototype.getJobs = function() {
 
 }
 
-function CSVToArray( strData, strDelimiter ){
-  strDelimiter = (strDelimiter || ",");
-  var objPattern = new RegExp(
-    (
-        // Delimiters.
-        "(\\" + strDelimiter + "|\\r?\\n|\\r|^)" +
-        // Quoted fields.
-        "(?:\"([^\"]*(?:\"\"[^\"]*)*)\"|" +
-        // Standard fields.
-        "([^\"\\" + strDelimiter + "\\r\\n]*))"
-    ),
-    "gi"
-  );
-  var arrData = [[]];
-  var arrMatches = null;
-  while (arrMatches = objPattern.exec( strData )){
-    var strMatchedDelimiter = arrMatches[ 1 ];
-    if (
-        strMatchedDelimiter.length &&
-        strMatchedDelimiter !== strDelimiter
-        ){
-        arrData.push( [] );
-    }
-    var strMatchedValue;
-    if (arrMatches[ 2 ]){
-        strMatchedValue = arrMatches[ 2 ].replace(
-            new RegExp( "\"\"", "g" ),
-            "\""
-            );
-
-    } else {
-        strMatchedValue = arrMatches[ 3 ];
-    }
-    arrData[ arrData.length - 1 ].push( strMatchedValue );
-  }
-  return( arrData );
-}
-
-GenericClusterInterface.prototype.getCompletedJobs = function(docs) {
-  var deferred = this.$q.defer();
-  var completedJobs = [];
-  for(var i = 0; i < docs.length; i++) {
-    // setup return document
-    var returnData = {
-      "_id": docs[i]._id
-    }
-    this.connectionService.runCommand("sacct -j " + docs[i].jobId + " -l --noconvert -P").then(function(data) {
-      // parse data and make list of docs to update the db with
-      var parsedData = CSVToArray(data, '|');
-      var headers = parsedData[0];
-      var jobData = parsedData[1];
-      var batchData = parsedData[2];
-      var desiredFields = [
-        "Elapsed",
-        "ReqMem",
-        "State",
-        "JobName"
-      ]
-      // loop through headers to get the indices of the desired fields
-      for(var j = 0; j < headers.length; j++) {
-        var index = desiredFields.indexOf(headers[j]);
-        if(index > -1) {
-          if(headers[j] == "JobName")
-            returnData[desiredFields[index]] = jobData[j] === undefined ? "" : jobData[j];
-          else
-            returnData[desiredFields[index]] = batchData[j] === undefined ? "" : batchData[j];
-        }
-      }
-      if(returnData.State != "COMPLETED") returnData = {}
-
-      completedJobs.push(returnData);
-      if(i==docs.length)
-        deferred.resolve(completedJobs);
-    });
-  }
-  return deferred.promise;
-}
-
 GenericClusterInterface.prototype.getStorageInfo = function() {
 
   // Return a promise if the storage info
