+++ HCCGo/package.json
@@ -23,7 +23,8 @@
     "node-notifier": "^4.6.0",
     "path": ">= 0.12.7",
     "ssh2": ">= 0.3.3",
-    "nedb": "^1.8.0"
+    "nedb": "^1.8.0",
+    "csv": "^1.1.0"
   },
   "devDependencies": {
     "devtron": "^1.2.1",
